Removing excess files that we won't be using. Adjusting type definitions as well.
